major redistricting simulation changes
