+++ Redistricting/cohesion.json
@@ -1,0 +1,99 @@
+{
+    "scenarios": [
+      {
+        "name": "fair_redistricting_cohesive",
+        "state_map_path": "/storage/home/nrk5343/work/MCMCAPSUHacks2025/pa_election_data.npy",
+        "flip_map": true,
+        "num_districts": 17,
+        "distribution_type": "fair",
+        "num_iterations": 20000,
+        "batch_size": 5000,
+        "use_parallel": true,
+        "pixels_per_move": 40,
+        "animation_frames": 150,
+        "animation_interval": 200,
+        "generate_heatmap": true,
+        "track_stats_evolution": true,
+        "anti_fragmentation": true,
+        "cohesion_phase": true,
+        "compactness_weight": 10.0,
+        "description": "Fair distribution with cohesive districts"
+      },
+      {
+        "name": "red_gerrymander_cohesive",
+        "state_map_path": "/storage/home/nrk5343/work/MCMCAPSUHacks2025/pa_election_data.npy",
+        "flip_map": true,
+        "num_districts": 17,
+        "distribution_type": "red_gerrymander",
+        "num_iterations": 20000,
+        "batch_size": 5000,
+        "use_parallel": true,
+        "pixels_per_move": 40,
+        "animation_frames": 150,
+        "animation_interval": 200,
+        "generate_heatmap": true,
+        "track_stats_evolution": true,
+        "anti_fragmentation": true,
+        "cohesion_phase": true,
+        "compactness_weight": 8.0,
+        "description": "Republican gerrymander with cohesive districts"
+      },
+      {
+        "name": "blue_gerrymander_cohesive",
+        "state_map_path": "/storage/home/nrk5343/work/MCMCAPSUHacks2025/pa_election_data.npy",
+        "flip_map": true,
+        "num_districts": 17,
+        "distribution_type": "blue_gerrymander",
+        "num_iterations": 20000,
+        "batch_size": 5000,
+        "use_parallel": true,
+        "pixels_per_move": 40,
+        "animation_frames": 150,
+        "animation_interval": 200,
+        "generate_heatmap": true,
+        "track_stats_evolution": true,
+        "anti_fragmentation": true,
+        "cohesion_phase": true,
+        "compactness_weight": 8.0,
+        "description": "Democratic gerrymander with cohesive districts"
+      },
+      {
+        "name": "incumbent_protection_cohesive",
+        "state_map_path": "/storage/home/nrk5343/work/MCMCAPSUHacks2025/pa_election_data.npy",
+        "flip_map": true,
+        "num_districts": 17,
+        "distribution_type": "incumbent",
+        "num_iterations": 20000,
+        "batch_size": 5000,
+        "use_parallel": true,
+        "pixels_per_move": 40,
+        "animation_frames": 150,
+        "animation_interval": 200,
+        "generate_heatmap": true,
+        "track_stats_evolution": true,
+        "anti_fragmentation": true,
+        "cohesion_phase": true,
+        "compactness_weight": 8.0,
+        "description": "Incumbent protection with cohesive districts"
+      },
+      {
+        "name": "compact_districts",
+        "state_map_path": "/storage/home/nrk5343/work/MCMCAPSUHacks2025/pa_election_data.npy",
+        "flip_map": true,
+        "num_districts": 17,
+        "distribution_type": "fair",
+        "num_iterations": 20000,
+        "batch_size": 5000,
+        "use_parallel": true,
+        "pixels_per_move": 40,
+        "animation_frames": 150,
+        "animation_interval": 200,
+        "generate_heatmap": true,
+        "track_stats_evolution": true,
+        "anti_fragmentation": true,
+        "cohesion_phase": true,
+        "compactness_weight": 15.0,
+        "description": "Extreme compactness optimization"
+      }
+    ]
+  }
