+++ Redistricting/gerrymander-scenarios-36023434.out
@@ -1,0 +1,12 @@
+Job started at Sun Mar 30 02:16:30 EDT 2025
+Running on node: p-mc-3475
+Available CPU cores: 40
+Memory:               total        used        free      shared  buff/cache   available
+Mem:          754Gi       4.6Gi       724Gi       6.3Gi        25Gi       739Gi
+Swap:            0B          0B          0B
+Submitted batch job 36023440
+Submitted batch job 36023441
+Submitted batch job 36023442
+Submitted batch job 36023443
+All scenario jobs submitted at Sun Mar 30 02:16:33 EDT 2025
+Check individual job logs for progress
